requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
ping3
APScheduler
